Fix: Resolve max_charging_energy not 0 when charger not occupied
Description:
When charger_occupancy of a charger is 0, i.e. there is no
vehicle connected to it at the current timestep,
max_charging_energy should be equal to 0. However,
if action for that charger is < 0, max_charging_energy
takes the value 10.

Add condition to check if charger is occupied at that timestep.
If occupied proceed as before, else max_charging_energy = 0.