Fix: Fix departure_times list incorrect size in station_simulation
Description:
In simulate_ev_charging_station(), when calculating departure_times,
if charger is occupied and also it is a timestep (i.e. hour) of
its departure, i.e. hour == departures[charger][vehicle] and
charger_occupancy[charger, hour] == 1, then it does not append
a departure to departure_times and it is then a list of size (27,)
instead of (28,).

Refactor hour < departures[charger][vehicle] to
hour <= departures[charger][vehicle].

Now departure_times is always of size (28,).