Refactor:
Description:
The if statement at the beggining of the __get_observations() method,
that checks if the single simulation cycle is completed, is only
accessed on the episode/epoch start and only after reset() mathod.
However, __get_observations() method is also called from step()
method aswell but it has no value from the beggining if statement
and checking the if condition is wasting time.

Extract beggining if statement to new method
__reset_variables_after_completed_day().

Now if statement is accessed only through reset() method.