Refactor: Remove obsolete energy['Consumed'] and functions that use it
Description:
energy['Consumed'] is declared in energy_calculations and is stored in
energy dictionary. It appears in actions_simulation originally and in
central_management_system because it is based on actions_simulation.
However, energy['Consumed'] never changes its state. It never affects
anything nor is used for anything. Therefore, it is wasting resources.

Remove energy['Consumed'].